Fixed final bug, now RMSE is meeting requirements. Issue was with the data type of delta_t_. I was using long long which would drop the decimals. Changed delta_t_ to double and then the issue was fixed.

+++ Term 2 - Sensor Fusion- Localization- and Control/Project 2 - Unscented Kalman Filters/src/ukf.h
@@ -74,7 +74,7 @@ public:
   long long previous_t_;
 
   // Delta timestamp
-  long long delta_t_;
+  double delta_t_;
 
   // R and H for Lidar update
   MatrixXd R_Lidar;
